refactor: Simplify `RegisterViewModel` and remove unused dependencies
- Extracted `RegisterUiState` into its own file for better modularity.
- Removed `UserRepository` dependency from `RegisterViewModel` and related methods.
- Simplified `AuthRepository` by eliminating `SessionManager` and redundant user creation logic after signup.
- Adjusted `RegisterScreen` and `LogInScreen` to align with the updated `AuthRepository` implementation.
- Upgraded AGP version in `libs.versions.toml` to `8.11.0`.
