+++ app/src/main/java/com/example/cookat/models/uiStates/RegisterUIState.kt
@@ -1,0 +1,9 @@
+package com.example.cookat.models.uiStates
+
+data class RegisterUiState(
+	val email: String = "",
+	val password: String = "",
+	val confirmPassword: String = "",
+	val isLoading: Boolean = false,
+	val error: String? = null
+)
