+++ app/src/main/java/com/example/cookat/repository/AuthRepository.kt
@@ -1,13 +1,11 @@
 package com.example.cookat.repository
 
-import com.example.cookat.data.local.session.SessionManager
 import com.example.cookat.data.remote.SupabaseClient
-import com.example.cookat.models.dbModels.users.UserModel
 import io.github.jan.supabase.auth.auth
 import io.github.jan.supabase.auth.providers.builtin.Email
 import io.github.jan.supabase.exceptions.BadRequestRestException
 
-class AuthRepository(private val sessionManager: SessionManager) {
+class AuthRepository {
 
 	private val auth = SupabaseClient.client.auth
 
@@ -17,11 +15,6 @@ class AuthRepository(private val sessionManager: SessionManager) {
 				this.email = email
 				this.password = password
 			}
-			val accessToken = auth.currentSessionOrNull()?.accessToken
-			if (accessToken != null) {
-				sessionManager.saveAccessToken(accessToken)
-			}
-			println("Bearer token after login: $accessToken")  // <-- acá lo imprimís
 			Result.success(Unit)
 		} catch (e: BadRequestRestException) {
 			Result.failure(Exception("Invalid email or password"))
@@ -30,7 +23,7 @@ class AuthRepository(private val sessionManager: SessionManager) {
 		}
 	}
 
-	suspend fun signUp(email: String, password: String, userRepository: UserRepository): Result<Unit> {
+	suspend fun signUp(email: String, password: String): Result<Unit> {
 		return try {
 			auth.signUpWith(Email) {
 				this.email = email
@@ -45,14 +38,14 @@ class AuthRepository(private val sessionManager: SessionManager) {
 
 			val accessToken = auth.currentSessionOrNull()?.accessToken
 			if (accessToken != null) {
-				sessionManager.saveAccessToken(accessToken)
+				//sessionManager.saveAccessToken(accessToken)
 				println("Bearer token after signUp: $accessToken")
 			}
 
 			val userId = getUserId()
 			if (userId != null) {
-				val user = UserModel(id = userId, email = email, username = "")
-				userRepository.createUser(user)
+				//val user = UserModel(id = userId, email = email, username = "")
+				//userRepository.createUser(user)
 			} else {
 				throw Exception("No se pudo obtener el ID del usuario luego del signUp")
 			}
